New Roact method for wrapping Roact components
A new `Roact` method allows a Roact component to be wrapped in a BasicState instace. This will allow direct access to any or specified keys within the BasicState store from the Roact component's own state object. You can modify the BasicState store as normal to re-render components with new information.